api/src/main/java/com/sirika/pymager/api/ImageId.java
/**
 * Copyright 2009 Sami Dalouche
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * 	http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.sirika.pymager.api;

import org.apache.commons.lang.StringUtils;
import org.apache.commons.lang.Validate;

/**
 * An Image ID
 * 
 * @author Sami Dalouche (sami.dalouche@gmail.com)
 * 
 */
public class ImageId {
    private String id;

    private ImageId(String id) {
        Validate.notNull(id);
        if ((!StringUtils.isAsciiPrintable(id))
                || (!StringUtils.isAlphanumeric(id))) {
            throw new IllegalArgumentException(
                    "ID cannot contain special characters");
        }
        this.id = id;
    }

    public static ImageId imageId(String id) {
        return new ImageId(id);
    }

    public String toString() {
        return id;
    }

    @Override
    public int hashCode() {
        final int prime = 31;
        int result = 1;
        result = prime * result + ((id == null) ? 0 : id.hashCode());
        return result;
    }

    @Override
    public boolean equals(Object obj) {
        if (this == obj)
            return true;
        if (obj == null)
            return false;
        if (getClass() != obj.getClass())
            return false;
        ImageId other = (ImageId) obj;
        if (id == null) {
            if (other.id != null)
                return false;
        } else if (!id.equals(other.id))
            return false;
        return true;
    }

}
